oracleasm uses a pseudo-filesystem to back its disk objects.  This allows it to use the inode hash to convert a bdev pointer to an actual disk object.
However, the test/set functions for comparing disk objects were not smart
enough.  A has collision on two different bdev pointers could cause an
incorrect lookup of a disk object.  A duplicate handle is returned.

We make a smarter test/set with a find_actor() function, very similar to
ocfs2.

This fixes Bug #6135457.



git-svn-id: http://oss.oracle.com/projects/oracleasm/src/trunk@380 af2e7e45-1c97-403c-8c3e-045742b29be9