oracleasm: slab build changes.
There have been a lot of changes to the slab code since we last worked
against mainline.

1) SLAB_CTOR_VERIFY disappeared.  The slab ctor is only called for real
   construction.
2) The slab dtor disappeared.  That's OK, we only passed NULL.
3) The slab ctor flags disappeared.  No one was using them.
4) The slab ctor cachep disappeared.  No one was using it.

This leaves us a lot of compatibility modes.  So, we check each in turn.

For dtor disappearing, we just need a compat handle for
kmem_cache_create.  Easy.

For the various versions of the ctor, we basically test and provide a
wrapper for each possibility.  The main oracleasm.c file just has the
modern one-argument version.



git-svn-id: http://oss.oracle.com/projects/oracleasm/src/trunk@396 af2e7e45-1c97-403c-8c3e-045742b29be9