oracleasm: New bio_end_io()
The kernel has changed bio_end_io().  It is now only called when a bio
is done, so if error is 0, the bio has fully completed (bi_size will be
zero and the original bi_size bytes will have been processed).  This
also means it's now void.

Let's do it that way.  Since we no longer have bytes_done, we need to
store the original io size on the asm_request object.  Then we can
continue where we left of.

We create a compat wrapper that handles the old bytes_done/bi_size for
older kernels.



git-svn-id: http://oss.oracle.com/projects/oracleasm/src/trunk@397 af2e7e45-1c97-403c-8c3e-045742b29be9