vendor: Add support for sles11 for header packaging
SLES11 builds kernel drivers on its own, but the headers are needed
to build the userspace.  We add the sles11 vendor to support everything,
including drivers.  Header packaging works, but driver building does not
due to kernel source issues; fix the kernel source, and the package will
work.

Along comes a couple cleanups for sles10.



git-svn-id: http://oss.oracle.com/projects/oracleasm/src/trunk@410 af2e7e45-1c97-403c-8c3e-045742b29be9