New substitution mechanism
We've previously used a combination of automake/autoconf and inline
shell/sed scripts to substitute variables in spec files and to control
version dependencies.

Leveraging the GNU Make Standard Library and a set of regexps we can
transition to a single, makefile-based approach to version dependencies
and substitutions.

Update vendor makefiles and rename .in files to -generic.

Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>