Introduce module parameter for block size selection
We have encountered a few devices which after a firmware update
communicate different characteristics to the OS. In particular, some
devices begin to report their physical block size. This in turn will
cause oracleasm to report a different block size to ASM and mounting the
disk group will fail.

Introduce a module parameter which permits the logical block size to be
reported instead of the physical.

Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>